voice_ingest/requirements.txt
aws_lambda_logging
mock>=4.0.2
python-dateutil>=2.8.1
botocore>=1.18.16
boto3>=1.15.16
python-dotenv>=0.14.0
elasticsearch_dsl
certifi
vaderSentiment
datefinder
pyyaml
lxml
marshmallow
marshmallow_dataclass
